pyproject.toml
[tool.poetry]
name = "bot_admins"
version = "0.1.0"
description = ""
authors = ["aldi <alkin.denis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.30"}
alembic = "^1.13.1"
aiogram = "^3.7.0"
pydantic-settings = "^2.2.1"
aiosqlite = "^0.20.0"
gspread-asyncio = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
